Projeto/power_outage_predictor/requirements.txt

mysql-connector
mysql-connector-python
mysql-connector-python-rf
numpy
pandas
openpyxl
singleton-decorator